+++ miden-crypto/src/ecdh/k256.rs
@@ -23,7 +23,7 @@ use crate::{
     ecdh::KeyAgreementScheme,
     utils::{
         ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable,
-        zeroize::{Zeroize, ZeroizeOnDrop},
+        zeroize::{Zeroize, ZeroizeOnDrop, Zeroizing},
     },
 };
 
@@ -116,9 +116,9 @@ impl EphemeralSecretKey {
         // is different than the one used in the `k256` one. This solution will no longer be needed
         // once `k256` gets a new release with a version of the `rand` dependency matching ours
         use k256::elliptic_curve::rand_core::SeedableRng;
-        let mut seed = [0_u8; 32];
-        rand::RngCore::fill_bytes(rng, &mut seed);
-        let mut rng = rand_hc::Hc128Rng::from_seed(seed);
+        let mut seed = Zeroizing::new([0_u8; 32]);
+        rand::RngCore::fill_bytes(rng, &mut *seed);
+        let mut rng = rand_hc::Hc128Rng::from_seed(*seed);
 
         let sk_e = k256::ecdh::EphemeralSecret::random(&mut rng);
         Self { inner: sk_e }